+++ pubmodules/trigger/rulesTrigger.js
@@ -20,10 +20,7 @@ var operatingHoursService = require("../../services/operatingHoursService");
 var sendMessageUtil = require("../../utils/sendMessageUtil");
 var cacheUtil = require("../../utils/cacheUtil");
 var cacheEnabler = require("../../services/cacheEnabler");
-
-
-
-const uuidv4 = require('uuid/v4');
+var UIDGenerator = require("../../utils/UIDGenerator");
 
 
 class RulesTrigger {
@@ -683,7 +680,7 @@ class RulesTrigger {
 
           
 
-          var request_id = 'support-group-'+id_project+"-"+uuidv4();
+          var request_id = 'support-group-'+id_project+"-"+UIDGenerator.generate();
           var id_user;
           var fullname;
           var email;